requirements.txt
psycopg[binary]==3.1.9
django==4.2.2
django-bootstrap5==23.3
djangorestframework==3.14.0
django-bootstrap-modal-forms==3.0.4
python-decouple==3.8